Documentation

All docstrings!

HORIZONSAPI.HORIZONSAPI — Module

Retrieve ephemeris data for free using JPL's REST API!

Warning

This package is not affiliated with or endorsed by NASA, JPL, Caltech, or any other organization! This is an independently written package by an astrodynamics hobbyist. See the HORIZONSAPI license file for more information about code sharing and usage!

Extended Help

License

MIT License

Copyright (c) 2022 Joe Carpinelli

Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

Imports

Base

Core

DocStringExtensions

Exports

HORIZONSAPI.REST — Module

The precise API for JPL's HORIZONS ephemeris platform, as described by the the official documentation.

Extended Help

Warning

This package is not affiliated with or endorsed by NASA, JPL, Caltech, or any other organization! This is an independently written package by an astrodynamics hobbyist. See the HORIZONSAPI license file for more information about code sharing and usage!

Imports

Base

Core

DocStringExtensions

HTTP

Exports

HORIZONSAPI.REST.CommonParameter — Type

struct CommonParameter <: HORIZONSAPI.REST.QueryParameter

A HORIZONS QueryParameter which is associated with many types of requests.

HORIZONSAPI Methods

CommonParameter(; name, default, allowable, description)

CommonParameter(name, default, allowable, description)

CommonParameter(pp; kws...)

CommonParameter(pp, di)

CommonParameter(pp, di)

HORIZONSAPI.REST.QueryParameter — Type

abstract type QueryParameter

An abstract supertype for all query parameters associated with HORIZONS API requests.

All QueryParameters subtypes must implement several methods:

REST.name

REST.defaults

REST.allowable

REST.description

Some subtypes must implement additional methods. For example, REST.EphemerisParameter must also implement the REST.ephemeristype method. The necessary methods for each subtype are described in each subtype's docstring.

HORIZONSAPI Methods

HORIZONSAPI.REST.allowable — Function

Returns the allowable values associated with a particular parameter. If no finite list of allowable values exists, then the singleton nothing is returned.

HORIZONSAPI.REST.defaults — Function

Returns the default values associated with a particular parameter. If no default values exist, then the singleton nothing is returned.

HORIZONSAPI.REST.description — Function

Returns the description associated with a particular parameter. If no description exists, then the singleton nothing is returned.

HORIZONSAPI.REST.ephemeristypes — Function

Returns the allowable ephemeris types (Observer, Vectors, or Elements) for a particular ephemeris-specific parameter.

HORIZONSAPI.REST.name — Function

Returns the name of a QueryParameter, as it would be written in a URL request.

HORIZONSAPI.REST.units — Function

Returns the units associated with a particular parameter. If no units exist, then the singeleton type nothing is returned.